Data Sets with Geometries of Dutch Provinces, Municipalities and Zipcodes

Data Sets with Geometries of Dutch Provinces, Municipalities and Zipcodes

gemeenten

ggdregios

jeugdregios

nuts3regios

postcodes4

provincies

veiligheidsregios

Format

A sf/data.frame with 3 variables:

...
column name of the identifier, which is the singular form of the name of the data set (i.e., gemeenten$gemeente)

area_km2
area in square kilometres

geometry
the polygonal form of the area

An object of class sf (inherits from data.frame) with 25 rows and 3 columns.

An object of class sf (inherits from data.frame) with 42 rows and 3 columns.

An object of class sf (inherits from data.frame) with 40 rows and 3 columns.

An object of class sf (inherits from data.frame) with 1060 rows and 3 columns.

An object of class sf (inherits from data.frame) with 12 rows and 3 columns.

An object of class sf (inherits from data.frame) with 25 rows and 3 columns.

Details

All data sets (maps) contain all of the Netherlands, except for postcodes4 which only contains the northern Netherlands.

See the repository file to update these data sets.